test internal broadcasting
need to make sure models are calculating things correctly as long
as axis 0 is the time axis.

This only applys when numpy is used internally, which is not a settable
option on all models. So also make a Base dummy method so it can at
least be called without error on all models.